+++ python/tests/test_yggdrasil/test_data/test_types/test_nested/test_struct/test_python.py
@@ -254,7 +254,7 @@ class TestWithFields:
 
         assert len(dtype.fields) == 1
         assert dtype.fields[0].name == "a"
-        assert dtype.fields[0].dtype.type_id == DataTypeId.INTEGER
+        assert dtype.fields[0].dtype.type_id == DataTypeId.INT64
 
 
 # ---------------------------------------------------------------------------
